refactor: rename many files to remove 'Card' from their names
