+++ application/controllers/proofController.js
