+++ application/controllers/spotPlateController.js
