+++ application/index.js
@@ -58,13 +58,13 @@ app.use('/cutting-setups', require('./controllers/cuttingSetupController'));
 app.use('/winding-setups', require('./controllers/windingSetupController'));
 app.use('/vendors', require('./controllers/vendorController'));
 app.use('/material-orders', require('./controllers/materialOrdersController'));
-app.use('/tickets', require('./controllers/ticketCardController'));
+app.use('/tickets', require('./controllers/ticketController'));
 app.use('/products', require('./controllers/productController'));
 app.use('/material-inventory', require('./controllers/materialInventoryController'));
 app.use('/hold-reasons', require('./controllers/holdReasonController'));
-app.use('/proofs', require('./controllers/proofCardController'));
-app.use('/die-lines', require('./controllers/dieLineCardController'));
-app.use('/spot-plates', require('./controllers/spotPlateCardController'));
+app.use('/proofs', require('./controllers/proofController'));
+app.use('/die-lines', require('./controllers/dieLineController'));
+app.use('/spot-plates', require('./controllers/spotPlateController'));
 
 databaseConnection.on('error', (error) => {
     throw new Error(`Error connecting to the database: ${error}`);
