+++ application/schemas/spotPlate.js
