+++ test/schemas/dieLine.spec.js
