+++ test/schemas/spotPlate.spec.js
